structured docker interface
